+++ CHANGELOG.md
@@ -6,7 +6,7 @@
 - Transcript: Add globally unique `uuid` field and `metadata` field to `Event`.
 - Transcript: Add `message_id` field to `ToolEvent` for corresponding `ChatMessageTool`.
 - Eval log: Add option to select sample by `uuid` in `read_eval_log_sample()`.
-- AgentSubmit: Add keep_submit_in_messages as an optional field.
+- ReAct agent: Add `keep_in_messages` option to `AgentSubmit` to preserve calls to `submit()` in message history.
 
 ## 0.3.112 (03 July 2025)